Move `Parser`, `ElementParser` and `PiParser` to the new module `parser`.
