+++ src/lib.rs
@@ -59,6 +59,7 @@ pub mod errors;
 pub mod escape;
 pub mod events;
 pub mod name;
+pub mod parser;
 pub mod reader;
 #[cfg(feature = "serialize")]
 pub mod se;
